fix #334: rename loader
